+++ src/Aggregate.ts
@@ -1,4 +1,0 @@
-export interface Aggregate<A> {
-  isEmpty: (a: A) => boolean
-  size: (a: A) => number
-}